+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 Changelog
 =========
 
+* Fixed a bug where the authentication refresh token flow did not update the token values
+
 v1.1.0 (2023-01-20)
 -------------------